a lot of refactoring

+++ NonTechTask/src/com/Dmitry_Elkin/PracticeTaskCRUD/repository/MyGenericRepositoryImpl.java
@@ -3,7 +3,6 @@ package com.Dmitry_Elkin.PracticeTaskCRUD.repository;
 import com.Dmitry_Elkin.PracticeTaskCRUD.model.BaseModelsMethsI;
 import com.Dmitry_Elkin.PracticeTaskCRUD.model.Specialty;
 import com.Dmitry_Elkin.PracticeTaskCRUD.model.Status;
-import com.Dmitry_Elkin.PracticeTaskCRUD.repository.GenericRepository;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 
@@ -23,9 +22,8 @@ import static java.nio.file.StandardCopyOption.REPLACE_EXISTING;
 /**
  *
  * attempt to make single Repository-class instead of SkillRepository, SpecialtyRepository and DeveloperRepository ))
- * @param <T>
+ * extends BaseModelsMethsI - чтобы JVM знал что у T есть соответствующие методы
  */
-//public class MyGenericRepositoryImpl<T> implements SpecialtyRepository {
 public class MyGenericRepositoryImpl<T extends BaseModelsMethsI> implements GenericRepository<T, Long> {
     final Class<T> typeParameterClass;
     private final String fileName;
@@ -52,10 +50,15 @@ public class MyGenericRepositoryImpl<T extends BaseModelsMethsI> implements Gene
     @Override
     public List<T> getAll(Status status) {
         List<T> itemList = new LinkedList<>();
+        if (!Files.exists(file)) {
+            System.out.println("file "+fileName+" not exist!");
+            return itemList;
+        }
+
         try {
             List<String> lines = Files.readAllLines(file);
             for (String jsonStr : lines) {
-                T item = (T) new Gson().fromJson(jsonStr, typeParameterClass);
+                T item = new Gson().fromJson(jsonStr, typeParameterClass);
                 if (status == null) {
                     itemList.add(item);
                 } else if (item.getStatus() == status) {
@@ -64,7 +67,8 @@ public class MyGenericRepositoryImpl<T extends BaseModelsMethsI> implements Gene
 
             }
         } catch (IOException e) {
-            throw new RuntimeException(e);
+//            throw new RuntimeException(e);
+            System.out.println("some io exception in module MyGenericRepositoryImpl in meth getAll: "+e.getMessage());
         }
         return itemList;
     }
@@ -79,7 +83,7 @@ public class MyGenericRepositoryImpl<T extends BaseModelsMethsI> implements Gene
     @Override
     public T getById(Long id) {
         if (!Files.exists(file)) {
-            System.out.println("The db-file is absent!");
+            System.out.println("The db-file "+file.getFileName()+" is absent!");
             return null;
         }
         String jsonStr;
@@ -92,7 +96,7 @@ public class MyGenericRepositoryImpl<T extends BaseModelsMethsI> implements Gene
                 }
             }
         } catch (IOException e) {
-            System.out.println("oops! some io exception was occurred "+e.getMessage());
+            System.out.println("some io exception in module MyGenericRepositoryImpl in meth getById: "+e.getMessage());
         }
 
         return null;
@@ -123,14 +127,19 @@ public class MyGenericRepositoryImpl<T extends BaseModelsMethsI> implements Gene
             Files.writeString(lastIdfile, "" + lastId, Charset.defaultCharset(), StandardOpenOption.CREATE);
 
         } catch (IOException e) {
-            System.out.println("oops, IO exception was occurred (( " + e.getMessage());
+            System.out.println("some io exception in module MyGenericRepositoryImpl in meth add: "+e.getMessage());
         }
     }
 
     public void update(T item){
+        if (!Files.exists(file)) {
+            System.out.println("file "+fileName+" not exist!");
+            return;
+        }
+
         try(
                 BufferedReader in = new BufferedReader(new FileReader(fileName));
-                BufferedWriter out = new BufferedWriter(new FileWriter(tmpFileName));
+                BufferedWriter out = new BufferedWriter(new FileWriter(tmpFileName))
                 )
         {
             String jsonStr;
@@ -146,15 +155,15 @@ public class MyGenericRepositoryImpl<T extends BaseModelsMethsI> implements Gene
                 out.newLine();
             }
         } catch (FileNotFoundException e) {
-            System.out.println("oops! File not found! "+e.getMessage());
+            System.out.println("oops! File "+fileName+" not found in module MyGenericRepositoryImpl in meth update: "+e.getMessage());
         } catch (IOException e) {
             //throw new RuntimeException(e);
-            System.out.println("oops! some IO exception : "+e.getMessage());
+            System.out.println("some io exception in module MyGenericRepositoryImpl in meth update: "+e.getMessage());
         }
         try {
             Files.move(tmpFile, file, REPLACE_EXISTING);
         } catch (IOException e) {
-            System.out.println("oops! some IO exception : "+e.getMessage());
+            System.out.println("some io exception in module MyGenericRepositoryImpl in meth update: "+e.getMessage());
         }
 
     }
